requirements.txt
fastapi[standard]
uvicorn
pydantic
python-dotenv
mysql-connector-python
werkzeug
python-jose
python-jose[cryptography] 
passlib[bcrypt]
python-decouple

